fix: AMP float16 Categorical crash + DRIVE_FOLDER_ID silent
Fix 1 – ppo_agent.py ActorNetwork.forward():
  AMP autocast produziert float16-logits. PyTorch Categorical
  validate_args schlaegt fehl mit IndependentConstraint(Real(),1)
  auch bei normalen Werten wegen float16-Praezision.
  logits.float() cast vor Categorical loest das ohne AMP-Performance
  zu verlieren (nur der Distributions-Aufbau ist float32, GRU bleibt f16).

Fix 2 – Notebook Zelle 0b:
  'DRIVE_FOLDER_ID nicht gesetzt' print entfernt.
  Funktion gibt jetzt still zurueck wenn Drive nicht konfiguriert.

+++ src/agents/ppo_agent.py
@@ -363,7 +363,10 @@ class ActorNetwork(BaseNetwork):
         self, state: torch.Tensor, hidden: Optional[Union[Tuple, torch.Tensor]] = None
     ) -> Tuple[Categorical, Optional[Union[Tuple, torch.Tensor]]]:
         logits, next_hidden = super().forward(state, hidden)
-        return Categorical(logits=logits), next_hidden
+        # Cast to float32: AMP autocast produziert float16-logits, Categorical
+        # validate_args schlaegt dann fehl (IndependentConstraint). float32 cast
+        # loest das ohne die AMP-Performance-Vorteile zu verlieren.
+        return Categorical(logits=logits.float()), next_hidden
 
 
 class CriticNetwork(BaseNetwork):
